Intermediate step where all executes are prepared for test debugging. Need to selectively make some optional now.

+++ spec/examples/record/list_spec.rb
@@ -6,7 +6,7 @@ describe Cequel::Record::List do
     key :permalink, :text
     column :title, :text
     list :tags, :text
-    list :contributor_ids, :int
+    list :contributor_ids, :bigint
   end
 
   let(:scope) { cequel[Post.table_name].where(:permalink => 'cequel') }
